Renamed q2-SpiecEasi as pretty-easi
