+++ pretty_easi/_network.py
