+++ pretty_easi/plugin_setup.py
@@ -9,9 +9,9 @@ from ._network import Network, NetworkDirectoryFormat, NetworkFormat
 
 
 plugin = Plugin(
-    name='spieceasi',
+    name='pretty-easi',
     version='0.0.0-dev',
-    website='https://github.com/BenKaehler/q2-SpiecEasi/tree/main/q2_SpiecEasi',  # noqa
+    website='https://github.com/BenKaehler/pretty-easi',
     package='q2_SpiecEasi',
     description=('This QIIME 2 plugin wraps SpiecEasi '
                  'and will be useful to anybody who wants to infer '
