+++ pretty_easi/tests/data/network.graphml
